feat: refacto and add remove
