+++ src/components/screens/RecentlyUsedApps.tsx
@@ -1,8 +1,9 @@
-import useRecentConnection from "@/hooks/useRecentConnection";
+import sortedRecentConnectionAppsAtom, {recentConnectionAppsAtom, initialValue as recentConnectionAppsInitialValue} from "@/atoms/recentConnectionAppsAtom";
 import AppCard from "../AppCard";
-import { Text } from "@ledgerhq/react-ui";
+import { Flex, Text } from "@ledgerhq/react-ui";
 import { t } from "i18next";
 import styled from "styled-components";
+import { useAtomValue,useSetAtom } from "jotai";
 
 const GridContainer = styled.div`
   width: 100%;
@@ -22,13 +23,21 @@ const StyledDiv = styled.div`
 `;
 
 export default function RecentlyUsedApps() {
-  const { lastConnectionApps } = useRecentConnection();
-
+  const lastConnectionApps = useAtomValue(sortedRecentConnectionAppsAtom);
+  const setLastConnectionApps = useSetAtom(recentConnectionAppsAtom)
+  
   return (
     <>
+
+    <Flex justifyContent={"space-between"}>
       <Text variant="extraSmall" color="neutral.c70" marginBottom={6}>
         {t("sessions.apps.lastConnection")}
       </Text>
+
+      <Text style={{cursor: "pointer"}} variant="extraSmall" color="neutral.c70" marginBottom={6} onClick={() => { setLastConnectionApps(recentConnectionAppsInitialValue)}}>
+        {t("sessions.apps.clearHistory")}
+      </Text>
+    </Flex>
 
       <GridContainer>
         {lastConnectionApps.map((app) => (
